postgres and mysql controllers cleaned up, organized, and functioning as intended. Working on making the addtional constraints property how das wants it
