+++ server/controllers/queries/mysql.queries.ts
@@ -1,0 +1,14 @@
+
+export const mysqlForeignKeyQuery = `
+  SELECT
+    COLUMN_NAME,
+    REFERENCED_TABLE_NAME,
+    REFERENCED_COLUMN_NAME,
+    CONSTRAINT_NAME,
+    TABLE_NAME
+  FROM
+    INFORMATION_SCHEMA.KEY_COLUMN_USAGE
+  WHERE
+    COLUMN_NAME = 'columnName'
+    AND REFERENCED_TABLE_NAME IS NOT NULL;
+    `
